WIP: Work around deleted instances when performing data migration
When upgrading from 2023.1 to 2023.2 we found that compute_id
data migrations took a significant period of time and/or didn't
complete correctly. This is due to deleted instances (without
a host/node set) could be re-evaluated in the loop every time it
executes as they never gain a compute_id.

I assume there is a good reason this approach wasn't used up
front, but having checked our databases it appears to be a valid
workaround in our deployments, so it is shared here in case it
helps others to move forwards.

Change-Id: I88c1e37a04a8c18490992d40526c40c63eb4ed11
Related-Bug: #2065403