Fix detaching devices by alias with mdevs
This change fixes the get_device_by_alias function
to properly handel devices that do not have an
alias in the domain xml.

By defintion if a device does not have an alais
it cant match a given alias, so this change adds
a hasattr check to ensure we do not attempt to check
the alias if its not defined.

This elimiants this entire class of bug for all devices
instead of just fixing it for a specific class.

Related-Bug: #1942345
Closes-Bug: #2074219
Change-Id: If417a43ea252647618e50391b63333f6b68bdfec