segments: select the correct physnet when it is defined by a port
Currently the selection of a physical_network is handled by
selecting the first one encountered in a segmented network. If a
multi-segment network uses multiple physnets and a port is already
assigned to one of these which is not the first in the list,
identification of the physnet_name will fail.

This patch adds the option to consult the port's assigned subnet
when looking up the physnet_name so that the correct segment can
be identified.

Change-Id: Ieda62057547b059ce63bb008e7441c499701e0b2